+++ app/src/main/java/com/cursocompose/notapp/domain/repository/NoteRepository.kt
@@ -4,7 +4,7 @@ import com.cursocompose.notapp.domain.model.Note
 
 interface NoteRepository {
     suspend fun getNotes(): List<Note>
-    suspend fun getNoteById(id: Int): Note?
+    suspend fun getNoteById(id: Long): Note?
     suspend fun insertNote(note: Note)
     suspend fun deleteNote(note: Note)
 }
